Espressoft-master/requirementsPip.txt
certifi
cffi
cryptography
dnspython
dpcpp-cpp-rt
idna
intel-cmplr-lib-rt
intel-cmplr-lic-rt
intel-opencl-rt
intel-openmp
mysql-connector-python
numpy
packaging
pandas
ply
protobuf
pycparser
PyQt5
PyQt5-Qt5
PyQt5-sip
pyqtgraph
python-dateutil
pytz
sip
six
tbb
toml
wincertstore
